SWITCHYARD-425: BeanSwitchYardScanner is not able to scan a project if the path contains a plus sign
https://issues.jboss.org/browse/SWITCHYARD-425
