+++ common/src/test/resources/classpathscan-test.jar
